edit: exercise/eliuds_eggs - size_t usage

+++ eliuds-eggs/eliuds_eggs.c
@@ -1,8 +1,9 @@
 #include "eliuds_eggs.h"
+#include <stddef.h>
 
 int egg_count(int display)
 {
-  int count = 0;
+  size_t count = 0;
   while (display != 0)
   {
     if (display & 1)
